pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supabase-llm-docs"
version = "1.0.0"
description = "Generate LLM-optimized documentation from Supabase SDK specifications"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Zyepher"}
]
keywords = ["supabase", "documentation", "llm", "generator", "openref", "ai", "claude", "gpt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-PyYAML",
    "types-requests",
]

[project.scripts]
supabase-llm-docs = "supabase_llm_docs.cli:main"

[project.urls]
Homepage = "https://github.com/Zyepher/supabase-llm-docs"
Repository = "https://github.com/Zyepher/supabase-llm-docs"
Issues = "https://github.com/Zyepher/supabase-llm-docs/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src/supabase_llm_docs --cov-report=html --cov-report=term"
